cluster/applications/packages/cortex/pyproject.toml
[tool.poetry]
name = "cortex"
version = "0.1.0"
description = ""
authors = ["Kai Aihara <kaidotio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"

openai = "^1.33.0"
tiktoken = "^0.7.0"
playwright = "^1.32.1"
beautifulsoup4 = "^4.12.2"
redis = "^4.6.0"
redis-om = "^0.2.1"
boto3 = "^1.26.162"
pillow = "^10.0.0"

httpx = "^0.25.1"

pydantic = "^1.10.10"
unstructured = {extras = ["pdf", "ppt", "pptx", "doc", "docx", "xlsx"], version = "^0.13"}
google-api-python-client = "^2.129.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
